chore(core): add and satisfy Clippy

+++ imageflow_core/src/flow/visualize.rs
@@ -42,21 +42,18 @@ pub fn notify_graph_changed(graph_ref: &mut Graph,
     if prev_graph_version >= 0 {
         let prev_filename =
             format!("job_{}_graph_version_{}.dot", r.debug_job_id, prev_graph_version);
-        match files_identical(&current_filename, &prev_filename)
+        if files_identical(&current_filename, &prev_filename)
             .expect(&format!("Comparison err'd for {} and {}", &current_filename, &prev_filename)) {
-            true => {
-                std::fs::remove_file(&current_filename).unwrap();
+            std::fs::remove_file(&current_filename).unwrap();
 
-                // Next time we will overwrite the duplicate graph. The last two graphs may
-                // remain dupes
-                Ok(None)
+            // Next time we will overwrite the duplicate graph. The last two graphs may
+            // remain dupes
+            Ok(None)
+        } else {
+            if r.render_graph_versions {
+                render_dotfile_to_png(&prev_filename)
             }
-            false => {
-                if r.render_graph_versions {
-                    render_dotfile_to_png(&prev_filename)
-                }
-                Ok(Some(GraphRecordingUpdate { next_graph_version: r.current_graph_version + 1 }))
-            }
+            Ok(Some(GraphRecordingUpdate { next_graph_version: r.current_graph_version + 1 }))
         }
     } else {
         Ok(None)
